+++ scanpipe/pipes/utilities.py
@@ -43,7 +43,7 @@ def analyze_licenses_and_sources(self):
 
 def tag_compliance_files(project):
     """
-    Tag compliance files.
+    Tag compliance files status for the provided `project`.
     """
     compliance_dirs = {
         "/licenses": "compliance-licenses",
@@ -57,6 +57,9 @@ def tag_compliance_files(project):
 
 
 def analyze_compliance_licenses(project):
+    """
+    Tag compliance licenses status for the provided `project`.
+    """
     qs = CodebaseResource.objects.project(project).status("compliance-licenses")
 
     for codebase_resource in qs: